Fix cZone background edit erasing the image in production
The background create/edit/delete endpoints each hand-counted `..`
segments off import.meta.url to locate the project root in production,
on the assumption a built route sits exactly as deep as its source
file under server/api/. Nitro's node-server preset actually nests it
three levels deeper (server/api/admin/backgrounds/[id]/... becomes
.output/server/chunks/routes/api/admin/backgrounds/_id/...), so every
one of those counts was short by 3.

mkdir(..., { recursive: true }) never surfaced the mistake -- it
silently created the wrong directory, the upload "succeeded" with a
200, and the path written to Background.imagePath (and copied onto
every CZone.background/layoutData referencing it) pointed at a file
that was never actually served. Editing a background looked like it
erased it because the DB now genuinely pointed nowhere.

Replaced the three hand-computed baseDir calculations with a shared
server/utils/backgroundStorage.js using process.cwd(), which is the
project root in both `nuxt dev` and production (PM2 launches
.output/server/index.mjs with the project root as cwd per
ecosystem.config.cjs) -- matching what the dev branch of this exact
code already did.
